requirements.txt
fastapi
uvicorn
supabase
pydantic
python-dotenv
psycopg2-binary
langgraph
langchain-google-genai
langchain-core
google-generativeai
httpx
